+++ src/utils/checkValidInput.js
