+++ src/utils/constant.js
@@ -1,4 +1,3 @@
-
 export const ERROR_MESSAGE = `1부터 9까지 서로 다른 수로 이루어진 3자리의 수를 입력해주세요!`;
 export const RETRY_ITEM = `
 <p><strong>🎉 정답을 맞추셨습니다! 🎉</strong></p>
